+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 - Removed duplicative Terminus::menu() in favor of Terminus\Helpers\Input::menu. (#768)
 - Moved Terminus::line() to Terminus\Outputters\Outputter. (#768)
 
+### Fixed
+- Missing creation dates in site data while using organizations site list command will no longer cause errors. (#766)
+
 ## [0.10.0] - 2015-12-15
 ### Added
 - New command `workflows show` displays details about a workflow (#687)