+++ CHANGELOG.md
@@ -10,7 +10,8 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 ### Fixed
 - Missing creation dates in site data while using organizations site list command will no longer cause errors. (#766)
 - Fixed headers in session token-based login. (#764)
-- `sites backups load --element=database` no longer errs upon calling the renamed function "backup". (#767)
+- `site backups load --element=database` no longer errs upon calling the renamed function "backup". (#767)
+- `site backups get --latest` bug wherein it was returning the oldest backup, rather than the most recent. (#770)
 
 ## [0.10.0] - 2015-12-15
 ### Added