+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 #Change Log
 All notable changes to this project starting with the 0.6.0 release will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org)
 
-## Master
+## [0.10.1] - 2015-01-12
 ### Added
 - `config/constants.yml` file to contain the default constants for Terminus. (#791)
 - Added a `--name=<regex>` filter to `sites list`. Use regex to filter the sites by name. (#802)